configure proper rendering of proj members after save, refactored components in folder
