+++ client/src/components/manageProjects/editProject.jsx
@@ -10,7 +10,7 @@ import EditIcon from '../../svg/Icon_Edit.svg?react';
 import PlusIcon from '../../svg/PlusIcon.svg?react';
 
 import { Typography, Box } from '@mui/material';
-import EditProjectMembers from './editProjectMembers';
+import EditProjectMembers from './editPMs/editProjectMembers';
 
 // Need to hold user state to check which type of user they are and conditionally render editing fields in this component
 // for user level block access to all except for the ones checked
